Implemented the MainWindowController for App Startup
The reason for this is because we want to start with a view that we created from scratch by calling it from the appDelegate DidFinishLaunching() method which instantiates an instance of the MainWindowController.swift class we create.  In the MainWindowController blueprint, it is a subclass of NSWindowController therefore we override the NibName property to point to our Xib file and then we add the properties for NSButton and NSTextField object instances.  We also implement the IBAction methods for when a user taps the NSButton objects on the view.  We'll need to connect these in a second.